backend/requirements.txt
fastapi
uvicorn
motor
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
aiohttp
beautifulsoup4
python-dotenv
openai>=1.30.0
